Isolate speech model libraries and split transcription execution
