+++ Sources/AudioQwen3ASRModel/Qwen3ASRAudioEncoder.swift
