+++ Sources/AudioSortformer/DiarizationOutput.swift
