+++ Sources/AudioSortformer/SortformerModel.swift
@@ -3,7 +3,6 @@
 import Foundation
 import MLX
 import MLXNN
-import MereRunCore
 
 // MARK: - FastConformer Encoder Components
 
@@ -551,223 +550,5 @@ public class SortformerModel: Module {
         let transEmbSeq = tfEncoder(embSeq, encoderMask: encoderMask)
         let preds = sortformerModules.forwardSpeakerSigmoids(transEmbSeq)
         return preds * encoderMask.expandedDimensions(axis: 2)
-    }
-
-    // MARK: - Offline Inference
-
-    public func generate(
-        audio: MLXArray,
-        sampleRate: Int = 16000,
-        threshold: Float = 0.5,
-        minDuration: Float = 0.0,
-        mergeGap: Float = 0.0
-    ) throws -> DiarizationOutput {
-        let startTime = Date()
-        let processor = config.processorConfig
-        guard sampleRate == processor.samplingRate else {
-            throw SortformerDiarizationError.unsupportedSampleRate(
-                actual: sampleRate,
-                expected: processor.samplingRate
-            )
-        }
-        guard audio.size > 0 else {
-            throw SortformerDiarizationError.emptyAudio
-        }
-
-        var waveform = audio.asType(.float32)
-        if waveform.ndim > 1 {
-            waveform = MLX.mean(waveform, axis: -1)
-        }
-
-        let (trimmed, trimOffset) = trimSilence(waveform, sampleRate: processor.samplingRate)
-        waveform = trimmed
-        let trimOffsetSeconds = Float(trimOffset) / Float(processor.samplingRate)
-        waveform = (1.0 / (MLX.abs(waveform).max() + 1e-3)) * waveform
-
-        let features = extractMelFeatures(
-            waveform,
-            sampleRate: processor.samplingRate,
-            nFft: processor.nFft,
-            hopLength: processor.hopLength,
-            winLength: processor.winLength,
-            nMels: processor.featureSize,
-            preemphasisCoeff: processor.preemphasis
-        )
-        let featureLengths = MLXArray([Int32(features.dim(2))])
-        let predictions = self(features, audioSignalLength: featureLengths)
-        eval(predictions)
-
-        let subsamplingFactor = config.fcEncoderConfig.subsamplingFactor
-        let frameDuration = Float(processor.hopLength * subsamplingFactor) / Float(processor.samplingRate)
-        var segments = Self.predsToSegments(
-            predictions[0],
-            frameDuration: frameDuration,
-            threshold: threshold,
-            minDuration: minDuration,
-            mergeGap: mergeGap
-        )
-
-        if trimOffset > 0 {
-            segments = segments.map {
-                DiarizationSegment(
-                    start: $0.start + trimOffsetSeconds,
-                    end: $0.end + trimOffsetSeconds,
-                    speaker: $0.speaker
-                )
-            }
-        }
-
-        return DiarizationOutput(
-            segments: segments,
-            numSpeakers: Set(segments.map(\.speaker)).count,
-            totalTime: Date().timeIntervalSince(startTime)
-        )
-    }
-    // MARK: - Postprocessing
-
-    public static func predsToSegments(
-        _ preds: MLXArray,
-        frameDuration: Float,
-        threshold: Float = 0.5,
-        minDuration: Float = 0.0,
-        mergeGap: Float = 0.0
-    ) -> [DiarizationSegment] {
-        let numFrames = preds.dim(0)
-        let numSpeakers = preds.dim(1)
-        var segments = [DiarizationSegment]()
-
-        // Single bulk GPU->CPU readback (row-major [frame, speaker]); all change
-        // detection runs in pure Swift to avoid per-frame .item() round-trips.
-        let flat = preds.asType(.float32).reshaped([-1]).asArray(Float.self)
-
-        for spk in 0..<numSpeakers {
-            var spkSegments = [DiarizationSegment]()
-            var segStart = -1
-            for f in 0..<numFrames {
-                let active = flat[f * numSpeakers + spk] > threshold
-                if active {
-                    if segStart < 0 { segStart = f }
-                } else if segStart >= 0 {
-                    let startTime = Float(segStart) * frameDuration
-                    let endTime = Float(f) * frameDuration
-                    if endTime - startTime >= minDuration {
-                        spkSegments.append(DiarizationSegment(start: startTime, end: endTime, speaker: spk))
-                    }
-                    segStart = -1
-                }
-            }
-            if segStart >= 0 {
-                let startTime = Float(segStart) * frameDuration
-                let endTime = Float(numFrames) * frameDuration
-                if endTime - startTime >= minDuration {
-                    spkSegments.append(DiarizationSegment(start: startTime, end: endTime, speaker: spk))
-                }
-            }
-
-            if mergeGap > 0 && spkSegments.count > 1 {
-                var merged = [spkSegments[0]]
-                for seg in spkSegments.dropFirst() {
-                    if seg.start - merged.last!.end <= mergeGap {
-                        merged[merged.count - 1] = DiarizationSegment(
-                            start: merged.last!.start, end: seg.end, speaker: seg.speaker
-                        )
-                    } else {
-                        merged.append(seg)
-                    }
-                }
-                spkSegments = merged
-            }
-
-            segments.append(contentsOf: spkSegments)
-        }
-
-        segments.sort { $0.start < $1.start }
-        return segments
-    }
-
-    // MARK: - Weight Sanitization & Loading
-
-    public static func sanitize(_ weights: [String: MLXArray]) -> [String: MLXArray] {
-        var sanitized = [String: MLXArray]()
-        let skipKeys: Set<String> = ["num_batches_tracked"]
-
-        let alreadyConverted = weights.keys.contains { $0.contains("subsampling.layers_") }
-
-        for (k, var v) in weights {
-            if skipKeys.contains(where: { k.contains($0) }) { continue }
-
-            var newK = k
-
-            if !alreadyConverted {
-                if newK.contains("fc_encoder.subsampling.layers.") {
-                    newK = newK.replacingOccurrences(of: "subsampling.layers.", with: "subsampling.layers_")
-                }
-
-                // Conv2d: PyTorch (O,I,H,W) → MLX (O,H,W,I)
-                if newK.contains("subsampling") && newK.contains("weight") && !newK.contains("linear") {
-                    if v.ndim == 4 {
-                        v = v.transposed(0, 2, 3, 1)
-                    }
-                }
-
-                // Conv1d: PyTorch (O,I,K) → MLX (O,K,I)
-                if (newK.contains("pointwise_conv1") || newK.contains("pointwise_conv2") || newK.contains("depthwise_conv"))
-                    && newK.contains("weight") {
-                    if v.ndim == 3 {
-                        v = v.transposed(0, 2, 1)
-                    }
-                }
-            }
-
-            sanitized[newK] = v
-        }
-
-        return sanitized
-    }
-
-    static func runtimeCompatibleWeights(
-        _ weights: [String: MLXArray],
-        promoteFloat16: Bool
-    ) -> [String: MLXArray] {
-        guard promoteFloat16 else { return weights }
-        return weights.mapValues { weight in
-            weight.dtype == .float16 ? weight.asType(.float32) : weight
-        }
-    }
-
-    public static func fromModelDirectory(_ modelURL: URL) throws -> SortformerModel {
-        // Load config
-        let configURL = modelURL.appendingPathComponent("config.json")
-        let configData = try Data(contentsOf: configURL)
-        let config = try JSONDecoder().decode(SortformerConfig.self, from: configData)
-
-        let model = SortformerModel(config)
-
-        // Load weights
-        let weightFiles = try FileManager.default.contentsOfDirectoryResolvingSymlinks(
-            at: modelURL, includingPropertiesForKeys: nil
-        ).filter { $0.pathExtension == "safetensors" }
-
-        var allWeights = [String: MLXArray]()
-        for file in weightFiles {
-            let weights = try loadArrays(url: file)
-            for (k, v) in weights {
-                allWeights[k] = v
-            }
-        }
-
-        let sanitized = sanitize(allWeights)
-        #if os(Linux) && arch(x86_64)
-        // mlx-swift cannot construct Float16 host scalars on x86_64. Promoting
-        // the compact Sortformer checkpoint keeps scalar model operations on
-        // the native CUDA path without triggering that host-only initializer.
-        let runtimeWeights = runtimeCompatibleWeights(sanitized, promoteFloat16: true)
-        #else
-        let runtimeWeights = sanitized
-        #endif
-        try model.update(parameters: ModuleParameters.unflattened(runtimeWeights), verify: .noUnusedKeys)
-        eval(model.parameters())
-
-        return model
     }
 }
