+++ Sources/MereRunCore/README.md
@@ -3,6 +3,7 @@
 Shared runtime validation, catalog assembly, generation protocols, and native
 runtime families. Model metadata and installed lookup live in `MereRunModelKit`;
 Core re-exports those types for existing callers.
+Shared full-attention caches live in `MereRunKVCache` and are also re-exported.
 
 - `Generation.swift`: image and chat request/response contracts.
 - `ImageGenerationOptions.swift` and `ImageGenerationPlan.swift`: typed image
