+++ Sources/MereRunCore/ZImageTurbo/Model/TextEncoder/LLMGeneration/README.md
@@ -3,7 +3,7 @@
 Shared MLX text-generation helpers used by ZImage text encoders and native chat
 engines that reuse the same primitives.
 
-- `KVCache.swift` owns the reusable full-attention cache protocol and
+- `MereRunKVCache` owns the reusable full-attention cache protocol and
   implementations. `KVCacheSimple` stores one scalar-offset row or equal-offset
   batches; `KVRaggedBatchCache` stores padded multi-row batches with per-row
   valid lengths so Qwen-family engines can prove variable-position decode
@@ -14,6 +14,7 @@ engines that reuse the same primitives.
 When adding a cache implementation, keep the protocol typed and split/fork
 semantics exact: a batched decode row must produce the same logits as the same
 row decoded independently.
+Core re-exports the cache library; Qwen ASR model layers depend on it directly.
 Cache forks use fresh array wrappers, not no-op dtype casts (which return the
 same object). Test fresh cache reads after either branch writes, rather than
 only comparing array views captured before those writes.
