+++ docs/repository-tour.md
@@ -26,11 +26,15 @@ mere-run/
 - `MereRunEvaluation`
 - `MereRunRelayKit`
 - `MereRunAdmission`
+- `MereRunResidency`
 - `MereRunModelKit`
 - `MereRunCore`
 - `AudioCore`
 - `AudioCodecs`
 - `AudioSTT`
+- `AudioQwen3ASRModel`
+- `AudioSortformer`
+- `MereRunKVCache`
 - `AudioTTS`
 - `MediaIO` (when its source target is available)
 - `mere.run` (the executable product backed by the `MereRunCLI` target)
@@ -147,11 +151,19 @@ on Apple media frameworks.
 
 ### `Sources/AudioSTT`
 
-Speech-to-text and speaker-diarization backends.
+Speech transcription orchestration, native executors, and compatibility exports.
 
-- `Qwen3ASR/`: native Qwen3 ASR path
+- `Qwen3ASR/`: Qwen3 model loading, tokenization, transcription, and streaming
 - `Parakeet/`: Parakeet transcription path
-- `Sortformer/`: native MLX Sortformer offline speaker diarization
+Public Qwen model and Sortformer types are re-exported from their owning libraries.
+
+### `Sources/AudioQwen3ASRModel`, `Sources/AudioSortformer`, and `Sources/MereRunKVCache`
+
+Qwen ASR configuration and model layers, the Sortformer diarization runtime,
+and shared full-attention caches build without `MereRunCore`. Qwen uses the
+same cache implementation as Core through `MereRunKVCache`. Sortformer uses
+`MereRunModelKit` for directory traversal. See
+[Speech runtime boundaries](./internals/speech-runtime-boundaries.md).
 
 ### `Sources/AudioTTS`
